// web/src/gen/model/platformDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ListOfString } from "./listOfString";

export interface PlatformDto {
  id: string;
  key: string;
  name: string;
  extensions: ListOfString;
  createdAt: string;
  /** @nullable */
  iconAssetId?: string | null;
  /** @nullable */
  iconUrl?: string | null;
}